oh boy we're back at it! Fixed up the team icons on the home screen
also fixed them up on the boxscore screen. That entire ui is still very sloppy. I need to fix it on individual plays as well but that's another api call to fight. Also updated the standings page to be the current year. That needs to be done with a dayjs function at some point but it works for now